requirements.txt
kafka-python==2.0.2
flask==2.2.2
pip~=21.3.1
wheel~=0.37.1
Jinja2~=3.1.2
setuptools~=60.2.0
requests~=2.28.1
pymongo~=4.3.2
datetime~=4.7
networkx==3.0
matplotlib
pandas
snscrape
transformers
scipy
detectlanguage
googletrans